tests: update test to test statment descriptor text updater action
